split tests to a separate file
